cleared up some todos
